+++ R/defaults.R
@@ -90,6 +90,7 @@ opts_chunk_attr = local({
   opts$fig.ext = as.list(unique(auto_exts))
   opts$external = opts$sanitize = NULL  # hide these two rare options
   opts$fig.process = 'function'
+  opts$R.options = 'list'
   opts
 })